refactor: double이 아닌 int를 반환하도록 수정

+++ src/main/java/domain/BlackJackGame.java
@@ -101,7 +101,7 @@ public class BlackJackGame {
     }
 
     //TODO: 메소드명 변경 고민
-    public double getResult(String name, int betMoney) {
+    public int getResult(String name, int betMoney) {
         Player player = players.findPlayerByName(name);
         BlackJackResult result = player.getResult(dealer);
         return result.calculatePrize(betMoney);
